Give BuilderHelper its own unit test
